chore: rename ResourceTable to ResourceViewList
