+++ packages/frontend/src/components/common/ResourceView/ResourceViewList/ResourceViewList.styles.ts
