+++ packages/frontend/src/pages/SavedDashboards.tsx
@@ -13,7 +13,6 @@ import {
     PageHeader,
 } from '../components/common/Page/Page.styles';
 import ResourceView from '../components/common/ResourceView';
-import { SortDirection } from '../components/common/ResourceView/ResourceTable';
 import {
     ResourceViewItemType,
     wrapResourceView,
@@ -24,6 +23,7 @@ import {
     ResourceEmptyStateIcon,
     ResourceTag,
 } from '../components/common/ResourceView/ResourceView.styles';
+import { SortDirection } from '../components/common/ResourceView/ResourceViewList';
 import { useCreateMutation } from '../hooks/dashboard/useDashboard';
 import { useDashboards } from '../hooks/dashboard/useDashboards';
 import { useSpaces } from '../hooks/useSpaces';
